Had to split up some functions from WiserItemsService into separate classes because I would otherwise have a circular reference. But it makes more sense anyway that there are separate classes for these functions.
