+++ GeeksCoreLibrary/Components/Account/Interfaces/IAccountsService.cs
@@ -3,6 +3,7 @@
 using System.Threading.Tasks;
 using GeeksCoreLibrary.Components.Account.Models;
 using GeeksCoreLibrary.Core.Models;
+using Constants = GeeksCoreLibrary.Components.Account.Models.Constants;
 
 namespace GeeksCoreLibrary.Components.Account.Interfaces
 {
