+++ GeeksCoreLibrary/Components/Account/Services/AccountsService.cs
@@ -19,6 +19,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Net.Http.Headers;
+using Constants = GeeksCoreLibrary.Components.Account.Models.Constants;
 
 namespace GeeksCoreLibrary.Components.Account.Services
 {
