+++ GeeksCoreLibrary/Modules/Databases/Services/MySqlDatabaseConnection.cs
@@ -25,6 +25,7 @@ using MySqlConnector;
 using Newtonsoft.Json;
 using Renci.SshNet;
 using ConnectionInfo = Renci.SshNet.ConnectionInfo;
+using Constants = GeeksCoreLibrary.Modules.Databases.Models.Constants;
 
 namespace GeeksCoreLibrary.Modules.Databases.Services
 {
