+++ GeeksCoreLibrary/Modules/Templates/Services/PagesService.cs
@@ -25,6 +25,7 @@ using GeeksCoreLibrary.Modules.Templates.ViewModels;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Constants = GeeksCoreLibrary.Modules.Templates.Models.Constants;
 
 namespace GeeksCoreLibrary.Modules.Templates.Services
 {
