+++ GeeksCoreLibrary/Modules/Templates/Services/TemplatesService.cs
@@ -34,6 +34,7 @@ using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json.Linq;
+using Constants = GeeksCoreLibrary.Modules.Templates.Models.Constants;
 using Template = GeeksCoreLibrary.Modules.Templates.Models.Template;
 
 namespace GeeksCoreLibrary.Modules.Templates.Services
